requirements.txt
flask~=3.1.0
flask_wtf
wtforms
mysql-connector-python